requirements.txt
asgiref==3.6.0
Django==4.2
mysqlclient==2.1.1
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
sqlparse==0.4.3
tzdata==2023.3
views==0.3
